// src/app/modules/provider/provider.interface.ts
export type TProvider = {
    name: string;
    about: Map<string, string>;
    specialists: string[];
    qualifications: string[];
    personal_info: Map<string, string>;
    professional_info: Map<string, string>;
    image: string;
    email: string;
    phone: string;
    x_url: string;
    facebook_url: string;
    instagram_url: string;
    linkedin_url: string;
    is_deleted: boolean;
};